chore: MetricQueryDataProvider and useMetricQueryDataContext
