+++ packages/frontend/src/components/MetricQueryData/MetricQueryDataProvider/index.tsx
@@ -1,17 +1,11 @@
 import {
     Explore,
-    formatItemValue,
-    getItemId,
-    hashFieldReference,
-    isDimension,
     ItemsMap,
     MetricQuery,
     PivotReference,
     ResultValue,
 } from '@lightdash/common';
-import { createContext, FC, useCallback, useContext, useState } from 'react';
-import { EChartSeries } from '../../hooks/echarts/useEchartsCartesianConfig';
-import { EchartSeriesClickEvent } from '../SimpleChart';
+import { createContext, FC, useCallback, useState } from 'react';
 
 export type UnderlyingDataConfig = {
     item: ItemsMap[string] | undefined;
@@ -27,7 +21,7 @@ export type DrillDownConfig = {
     pivotReference?: PivotReference;
 };
 
-type MetricQueryDataContext = {
+export type MetricQueryDataContext = {
     tableName: string;
     explore: Explore | undefined;
     metricQuery?: MetricQuery;
@@ -43,55 +37,9 @@ type MetricQueryDataContext = {
     closeDrillDownModal: () => void;
 };
 
-export const getDataFromChartClick = (
-    e: EchartSeriesClickEvent,
-    itemsMap: ItemsMap,
-    series: EChartSeries[],
-): UnderlyingDataConfig => {
-    const pivotReference = series[e.seriesIndex]?.pivotReference;
-    const selectedFields = Object.values(itemsMap).filter((item) => {
-        if (
-            !isDimension(item) &&
-            pivotReference &&
-            pivotReference.field === getItemId(item)
-        ) {
-            return e.dimensionNames.includes(
-                hashFieldReference(pivotReference),
-            );
-        }
-        return e.dimensionNames.includes(getItemId(item));
-    });
-    const selectedMetricsAndTableCalculations = selectedFields.filter(
-        (item) => !isDimension(item),
-    );
-
-    let selectedField: ItemsMap[string] | undefined = undefined;
-    if (selectedMetricsAndTableCalculations.length > 0) {
-        selectedField = selectedMetricsAndTableCalculations[0];
-    } else if (selectedFields.length > 0) {
-        selectedField = selectedFields[0];
-    }
-    const selectedValue = selectedField
-        ? e.data[getItemId(selectedField)]
-        : undefined;
-    const fieldValues: Record<string, ResultValue> = Object.entries(
-        e.data,
-    ).reduce((acc, entry) => {
-        const [key, val] = entry;
-        return { ...acc, [key]: { raw: val, formatted: val } };
-    }, {});
-
-    return {
-        item: selectedField,
-        value: {
-            raw: selectedValue,
-            formatted: formatItemValue(selectedField, selectedValue),
-        },
-        fieldValues,
-        pivotReference,
-    };
-};
-const Context = createContext<MetricQueryDataContext | undefined>(undefined);
+export const Context = createContext<MetricQueryDataContext | undefined>(
+    undefined,
+);
 
 type Props = {
     tableName: string;
@@ -155,21 +103,5 @@ const MetricQueryDataProvider: FC<Props> = ({
         </Context.Provider>
     );
 };
-
-export function useMetricQueryDataContext<S extends boolean = false>(
-    failSilently?: S,
-): S extends false
-    ? MetricQueryDataContext
-    : MetricQueryDataContext | undefined {
-    const context = useContext(Context);
-
-    if (context === undefined && failSilently !== true) {
-        throw new Error(
-            'useMetricQueryDataContext must be used within a UnderlyingDataProvider',
-        );
-    }
-
-    return context!;
-}
 
 export default MetricQueryDataProvider;
